[gui] remember window size per each resolution
